requirements.txt
faiss-cpu
langchain
langchain-community
langchain-ollama
streamlit
pypdf
sentence_transformers